# swagger/openapi.yaml
openapi: 3.1.0

info:
  title: MLforge API
  version: 0.1.0
  description: |
    A lightweight, language-independent experiment tracking API.

    MLforge manages experiments, runs, parameters, metrics, and plots
    metadata for machine learning and other computational workloads.

tags:
  - name: System
  - name: Experiments
  - name: Runs
  - name: MetricPlots

paths:
  /health:
    get:
      tags:
        - System
      operationId: getHealth
      summary: Check server health
      responses:
        "200":
          description: Server is healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"

  /api/experiments:
    get:
      tags:
        - Experiments
      operationId: listExperiments
      summary: List experiments
      responses:
        "200":
          description: Experiment list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExperimentSummariesResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"

    post:
      tags:
        - Experiments
      operationId: createExperiment
      summary: Create an experiment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExperimentRequest"
      responses:
        "200":
          description: Experiment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExperimentIdResponse"
        "409":
          description: Experiment name already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        default:
          $ref: "#/components/responses/ErrorResponse"

  /api/experiments/{experimentId}:
    parameters:
      - $ref: "#/components/parameters/ExperimentId"

    get:
      tags:
        - Experiments
      operationId: getExperiment
      summary: Get an experiment
      responses:
        "200":
          description: Experiment details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExperimentDetailsResponse"
        "404":
          $ref: "#/components/responses/NotFoundResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"

    patch:
      tags:
        - Experiments
      operationId: updateExperiment
      summary: Update an experiment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExperimentRequest"
      responses:
        "204":
          description: Experiment updated
        "404":
          $ref: "#/components/responses/NotFoundResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"

    delete:
      tags:
        - Experiments
      operationId: deleteExperiment
      summary: Delete an experiment
      description: |
        Deletes the experiment and its associated runs.
      responses:
        "204":
          description: Experiment deleted
        "404":
          $ref: "#/components/responses/NotFoundResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  
  /api/experiments/{experimentId}/runs:
    parameters:
      - $ref: "#/components/parameters/ExperimentId"

    post:
      tags:
        - Runs
      operationId: createRun
      summary: Create and start a run
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RunRequest"
      responses:
        "200":
          description: Run created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunIdResponse"
        "404":
          $ref: "#/components/responses/NotFoundResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"

  /api/runs/{runId}:
    parameters:
      - $ref: "#/components/parameters/RunId"
    
    delete:
      tags:
        - Runs
      operationId: deleteRun
      summary: Delete a run
      description: |
        Deletes the run and its associated parameters and metrics.
      responses:
        "204":
          description: Run deleted
        "404":
          $ref: "#/components/responses/NotFoundResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"

  /api/runs/{runId}/finish:
    parameters:
      - $ref: "#/components/parameters/RunId"

    put:
      tags:
        - Runs
      operationId: finishRun
      summary: Mark a run as completed
      responses:
        "204":
          description: Run completed
        "404":
          $ref: "#/components/responses/NotFoundResponse"
        "409":
          description: Invalid run status transition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        default:
          $ref: "#/components/responses/ErrorResponse"

  /api/runs/{runId}/fail:
    parameters:
      - $ref: "#/components/parameters/RunId"

    put:
      tags:
        - Runs
      operationId: failRun
      summary: Mark a run as failed
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FailRunRequest"
      responses:
        "204":
          description: Run marked as failed
        "404":
          $ref: "#/components/responses/NotFoundResponse"
        "409":
          description: Invalid run status transition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        default:
          $ref: "#/components/responses/ErrorResponse"

  /api/metrics/{metricId}/plots:
    parameters:
      - $ref: "#/components/parameters/MetricId"

    post:
      tags:
        - MetricPlots
      operationId: postMetricPlots
      summary: Post metric plots
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMetricPlotsRequest"
      responses:
        "204":
          description: Metric updated
        "404":
          $ref: "#/components/responses/NotFoundResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  
  /api/plots:

    post:
      tags:
        - MetricPlots
      operationId: getMetricPlots
      summary: Get metric plots for an experiment
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GetMetricPlotsRequest"
      responses:
        "200":
          description: Metric plots retrieved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetMetricPlotsResponse"
        "404":
          $ref: "#/components/responses/NotFoundResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"

components:
  parameters:
    ExperimentId:
      name: experimentId
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1

    RunId:
      name: runId
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1

    MetricId:
      name: metricId
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1

  responses:
    ErrorResponse:
      description: Request failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"

    NotFoundResponse:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"

  schemas:
    HealthResponse:
      type: object
      required:
        - status
        - version
      properties:
        status:
          type: string
          enum:
            - ok
        version:
          type: string
          example: 0.1.0

    RunStatus:
      type: integer
      enum:
        - 0 # running
        - 1 # completed
        - 2 # failed

    Parameter:
      type: object
      additionalProperties: false
      required:
        - id
        - name
        - value
      properties:
        id:
          type: integer
          format: int64
          minimum: 1
        name:
          type: string
          maxLength: 32
        value:
          type: string
          maxLength: 32
    
    MetricPlotRange:
      type: object
      additionalProperties: false
      properties:
        startStep:
          type: integer
          format: int64
          minimum: 0
          default: 0
        endStep:
          type: integer
          format: int64
          minimum: 1
          default: 1000
    
    Metric:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: integer
          format: int64
          minimum: 1
        name:
          type: string
          minLength: 1
          maxLength: 32
    
    RunSummary:
      type: object
      required:
        - id
        - name
        - description
        - failureReason
        - status
        - startedAt
        - endedAt
      properties:
        id:
          type: integer
          format: int64
          minimum: 1
        name:
          type:
            - string
            - "null"
          maxLength: 32
        description:
          type:
            - string
            - "null"
          maxLength: 512
        failureReason:
          type:
            - string
            - "null"
          maxLength: 512
        status:
          $ref: "#/components/schemas/RunStatus"
        startedAt:
          type: string
          format: date-time
        endedAt:
          type:
            - string
            - "null"
          format: date-time
    
    RunDetails:
      type: object
      required:
        - runSummary
        - parameters
        - metricSummaries
      properties:
        runSummary:
          $ref: "#/components/schemas/RunSummary"
        parameters:
          type: array
          items:
            $ref: "#/components/schemas/Parameter"
        metrics:
          type: array
          items:
            $ref: "#/components/schemas/Metric"

    MetricPlots:
      type: object
      required:
        - metricId
        - plots
      properties:
        metricId:
          type: integer
          format: int64
          minimum: 1
        plots:
          type: object
          propertyNames:
            pattern: '^[1-9][0-9]*$'
          additionalProperties:
            type: number
            format: double
    
    RunMetricPlots:
      type: object
      required:
        - runId
        - metricId
        - plots
      properties:
        runId:
          type: integer
          format: int64
          minimum: 1
        metricPlots:
          type: array
          items:
            $ref: "#/components/schemas/MetricPlots"
    
    ExperimentSummary:
      type: object
      required:
        - id
        - name
        - description
        - createdAt
      properties:
        id:
          type: integer
          format: int64
          minimum: 1
        name:
          type: string
          minLength: 1
          maxLength: 32
        description:
          type: string
          maxLength: 512
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type:
            - string
            - "null"
          format: date-time
    
    ExperimentDetails:
      type: object
      required:
        - experimentSummary
        - runDetails
      properties:
        experimentSummary:
          $ref: "#/components/schemas/ExperimentSummary"
        runDetails:
          type: array
          items:
            $ref: "#/components/schemas/RunDetails"

    ExperimentRequest:
      type: object
      additionalProperties: false
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 32
        description:
          type: string
          maxLength: 512
    
    ExperimentIdResponse:
      type: object
      required:
        - experimentId
      properties:
        id:
          type: integer
          format: int64
          minimum: 1

    ExperimentSummariesResponse:
      type: object
      required:
        - experimentSummaries
      properties:
        experimentSummaries:
          type: array
          items:
            $ref: "#/components/schemas/ExperimentSummary"
    
    ExperimentDetailsResponse:
      type: object
      required:
        - experimentDetails
      properties:
        experimentDetails:
          $ref: "#/components/schemas/ExperimentDetails"

    RunRequest:
      type: object
      additionalProperties: false
      properties:
        name:
          type: string
          maxLength: 32
        description:
          type: string
          maxLength: 512
        parameters:
          type: object
          additionalProperties:
            type: string
        metrics:
          type: array
          items:
            type: string
    
    RunIdResponse:
      type: object
      required:
        - experimentId
      properties:
        id:
          type: integer
          format: int64
          minimum: 1
        metricIds:
          type: array
          items:
            type: integer
            format: int64
            minimum: 1

    FailRunRequest:
      type: object
      additionalProperties: false
      properties:
        failureReason:
          type: string
          maxLength: 512

    CreateMetricPlotsRequest:
      type: object
      additionalProperties: false
      required:
        - plots
      properties:
        plots:
          type: object
          propertyNames:
            pattern: '^[1-9][0-9]*$'
          additionalProperties:
            type: number
            format: double

    GetMetricPlotsRequest:
      type: object
      additionalProperties: false
      required:
        - plotRange
        - runIds
      properties:
        plotRange:
          $ref: "#/components/schemas/MetricPlotRange"
        runIds:
          type: array
          items:
            type: integer
            format: int64
            minimum: 1
    
    GetMetricPlotsResponse:
      type: object
      required:
        - runMetricPlots
      properties:
        runMetricPlots:
          type: array
          items:
            $ref: "#/components/schemas/RunMetricPlots"

    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: RESOURCE_NOT_FOUND
        message:
          type: string
          example: The requested run was not found.
        details:
          type:
            - object
            - "null"
          additionalProperties: true